refactor: Swap inspect and uki modules hierarchy
Since we can still afford changing the api, we should use the noun-first
command conventions for the tool as it scales better with time.
