+++ lowell-core/src/lib.rs
@@ -1,3 +1,3 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 pub mod formats;
-pub mod inspect;
+pub mod uki;
