+++ lowell-core/src/uki/mod.rs
@@ -1,3 +1,3 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 pub mod ext;
-pub mod uki;
+pub mod inspect;
